+++ src/index.js
@@ -5,7 +5,7 @@ import './style.scss';
 
 import App from './App.jsx';
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
+const root = ReactDOM.createRoot(document.querySelector('#root'));
 
 root.render(
    <React.StrictMode>